[core] Add two symbol name of `ProcessProfilingInfo`
Android 12 changes the symbol of `ProcessProfilingInfo`, just follow the change.
MIUI added a new function and move the implementation to it, the old function just call the new one, cannot be hooked due to small size, just hook the new one instead. Also removed the previous workaround as we really fixed this issue. This fixes #18.