Update Android Studio
NDK is stuck at r25.2 because we want to keep Android 4.4 support.

Before, targetSdk was intentionally set to 27 because more strict SELinux rules apply to apps with targetSdk >= 28, which prevent me from making crash tombstones world-readable. When I wrote the code, I only had a old Android Lolipop device with a broken USB tail plug, and when I connected it to PC, it always disconnects immediately, making it almost impossible for me to program. At that time, the only way to debug my project was to make the crash tombstones world-readable so I can use a file manager to view them and find the cause。。。You can imagine how helpless a junior high school student without any money feels。。。
To make sure we can find if any new changes affect us on time, set it to the latest Android version.