[core] Fix unpredictable usage of STREX
Instruction format: STREX<c> <Rd>, <Rt>, [<Rn>{, #imm}]

However, this instruction requires (Rd != Rt && Rd != Rn), otherwise its behavior is unpredictable. Check ARM DDI 0406C.b, A8.8.213. for more info. (What an amazing design, love come from China...)

This means if we want to use this instruction, we must find 3 registers available. We know that R0 and IP (R12) are available, so we need to find another register for use. In AAPCS/ATPCS, R0-R3 are caller-save registers, but R1-R3 is used to pass arguments. If we use R1-R3, the arguments will be clobbered. R4-R11 are callee-save, IP (R12) is already used, and making use of R13-R15 is impossible and also unpredictable.
After the above analysis, we can determine that we must choose a register in R4-R11 to use. In normal programs, the usual approach is to push the register we use onto stack and later restore it from stack. But in art, I'm not sure if changing SP will affect CC (Concurrect Copying Garbage Collector). In ART, R4 is neither managed callee-save, nor argument register, althrough it is native callee-save but the value coming from managed code can be clobbered. But I'm also not sure if clobbering it is safe. Anywhy, better safe than sorry.
How can we backup a register without pushing it onto the stack? Well, the stack is allocated when the thread being created, and when allocating space on the stack, we just decrease SP by the size we allocate, and then [SP+size]~[SP] is available for use. Because the stack grows downward, [STACK END]~[SP] can never be occupied. We (ab)use the un-allocated space on the stack to store R4. Because we do not call any function, and ART always call sigaltstack() that makes sure signal handlers use another stack, no function calls will happen between storing and restoring, so it's safe to use un-allocated stack space.

Fix #36, fix #37