[core] Exclude unwanted bits when parsing boolean
When getting arguments from the stack, we take every argument as uint64_t which is 8 bytes, thus the value may contain unneeded bits belonging to the next argument. We already exclude unwanted bits when parsing other types of parameters, but we forgot to do so for boolean whose size is 4 bytes (aligned upward from 1 bit). Also add a new test to cover this edge case.

Fix #63